feat: add Outlet
