+++ src/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
 import './assets/fonts/fonts.css';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
+import { Layout } from './components/Layout/Layout';
 import Home from './pages/Home/Home';
 import NotFound from './pages/NotFound/NotFound';
 
@@ -11,8 +12,10 @@ function App() {
       <ErrorBoundary>
         <BrowserRouter>
           <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="*" element={<NotFound />} />
+            <Route path="/" element={<Layout />}>
+              <Route index element={<Home />} />
+              <Route path="*" element={<NotFound />} />
+            </Route>
           </Routes>
         </BrowserRouter>
       </ErrorBoundary>
