+++ src/components/Catalog/Catalog.module.scss
@@ -1,6 +1,4 @@
-main {
-  margin-bottom: 2rem;
-}
+
 
 .searchContainer {
   display: flex;
