+++ src/components/Header/Header.module.scss
@@ -1,9 +1,9 @@
 .header {
-  height: 2rem;
+  margin-bottom: 1rem;
 }
 
 .title {
   font-family: 'rickandmorty';
-  font-size: 3rem;
+  font-size: 3em;
   color: #00b1c8;
 }
