+++ src/components/Layout/Layout.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Outlet } from 'react-router-dom';
+import Footer from '../Footer/Footer';
+import Header from '../Header/Header';
+import style from './Layout.module.scss';
+
+interface Props {
+  className?: string;
+}
+
+export const Layout: React.FC<Props> = () => {
+  return (
+    <>
+      <Header />
+      <main className={style.main}>
+        <Outlet />
+      </main>
+      <Footer />
+    </>
+  );
+};
