+++ src/pages/Home/Home.tsx
@@ -1,14 +1,10 @@
 import { FC } from 'react';
-import Footer from '../../components/Footer/Footer';
-import Header from '../../components/Header/Header';
-import Main from '../../components/Main/Main';
+import Catalog from '../../components/Catalog/Catalog';
 
 const Home: FC = () => {
   return (
     <>
-      <Header />
-      <Main />
-      <Footer />
+      <Catalog />
     </>
   );
 };
