+++ src/pages/NotFound/NotFound.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import styles from './NotFound.module.scss';
 
 const NotFound: React.FC = () => {
@@ -9,6 +10,7 @@ const NotFound: React.FC = () => {
       <p className={styles.message}>
         Oops! The page you are looking for does not exist.
       </p>
+      <Link to="/">Go back to Home</Link>
     </div>
   );
 };
